rename word_counter
