--- tests/compare_tables/darwin_show_result_no_list.txt
@@ -1,6 +1,0 @@
-                                                                                                                                                                                                          
-
-   Found 2 file(s).
-   Total combined size: 49.0 B.
-   Average file size: 24.0 B (max: 49.0 B, min: 0.0 B).
-
