+++ tests/performance_tests/cprofile_test.py
@@ -17,7 +17,7 @@ import cProfile
 import os
 import sys
 
-from count_files.utils.word_counter import show_result_for_search_files, show_2columns
+from count_files.utils.viewing_modes import show_result_for_search_files, show_2columns
 from count_files.utils.file_handlers import search_files, count_files_by_extension
 from count_files.__main__ import main_flow
